Added Database Class
added extractDataToStringMap Method and getDataOnce Method
